Update structure and move forms into separate folder.
